requirements.txt
google-cloud-bigquery~=3.34.0
requests~=2.32.4